backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.23
alembic>=1.13.1
asyncpg>=0.29.0

# Caching & Queues
redis>=5.0.1
kafka-python>=2.0.2

# ML/AI
xgboost>=1.7.6
scikit-learn>=1.3.0
pandas>=1.5.3
numpy>=1.24.3
lightgbm>=3.3.5
catboost>=1.2
shap>=0.42.0
lime>=0.2.0.1
mlflow>=2.7.1

# Feature Store
feast

# LLM & Reasoning
langchain>=0.0.354
ollama>=0.2.1

# Monitoring & Observability
prometheus-client>=0.19.0
structlog>=23.2.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Async & Concurrency
asyncio-mqtt>=0.16.1
aiofiles>=23.2.1

# HTTP & Web
httpx>=0.27.0
websockets>=12.0

# Data Processing
pyarrow>=13.0.0
joblib>=1.3.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Code Quality
black
isort
flake8
bandit
safety

# Development
jupyter>=1.0.0
ipykernel>=6.27.1
matplotlib>=3.8.2
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
tqdm>=4.66.1

# Added for runtime support
tabpfn
slowapi
prometheus-fastapi-instrumentator
PyJWT
requests
river
langchain-community